+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `dep run` command [#1263]
 - Added new `-o` option which allow to override default configuration
 - Added `dep autocomplete` command
+- Added `dep config:hosts` task to show inventory
 
 ### Changed
 - Use either one of `command`, `which` or `type` commands to locate custom binary path.